pyproject.toml
[tool.poetry]
name = "projeto-final-es"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "projeto_final_es"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.3"
sqlalchemy-json = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
